better contiguous_count

+++ tests/kernels/contiguous_count_test.py
@@ -35,6 +35,6 @@ class ContiguousCountTest(TestCommons):
         x = torch.randint(0, _MAX_EXPERTS, (size,), device=device, dtype=torch.long)
 
         z_kernel = function(x=x, size=_MAX_EXPERTS, kernel_backend=kernel_backend, BLOCK_SIZE_B=BLOCK_SIZE_B)
-        z_expected = x.view(-1).bincount(minlength=_MAX_EXPERTS)
+        z_expected = x.view(-1).bincount(minlength=_MAX_EXPERTS).to(torch.uint32)
 
         self.assert_equal_tensors(z_kernel, z_expected, True)
